lite: Swapped UART and GPIO pin header rows in both schematic and PCB.
This allows to more easily use a mezzanine board.

It is now possible to solder a single row vertical pin header on the row
that is located inside the PCB and a right-angled single row pin header on
the row facing the PCB edge to still get access to the UART pins.
This way, the GPIOs can reach the mezzanine board and there is no need to
"forward" the UART signals to the mezzanine board.

Signed-off-by: Adrien Ricciardi <aricciardi@baylibre.com>